requirements.txt
streamlit
fpdf2
docx2pdf 